+++ src/main/java/cn/infinivision/dataforce/busybee/pb/rpc/BMRangeRequestOrBuilder.java
@@ -13,9 +13,9 @@ public interface BMRangeRequestOrBuilder extends
   com.google.protobuf.ByteString getKey();
 
   /**
-   * <code>uint32 start = 2;</code>
+   * <code>uint64 start = 2;</code>
    */
-  int getStart();
+  long getStart();
 
   /**
    * <code>uint64 limit = 3;</code>
